Remove AdvantageKit
