+++ src/main/java/frc3512/robot/Constants.java
@@ -5,60 +5,21 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
 import edu.wpi.first.math.util.Units;
-import edu.wpi.first.wpilibj.RobotBase;
 import frc3512.lib.config.SwerveModuleConstants;
 
 /** Constants for the robot project */
 public final class Constants {
 
-  /** Robot mode enum. */
-  public static enum RobotMode {
-    ROBOT_2022_REAL,
-    ROBOT_2022_SIM
-  }
-
-  /** Running mode enum. */
-  public static enum RunningMode {
-    REAL,
-    REPLAY,
-    SIM
-  }
-
   /** General robot constants */
   public static final class General {
-    /** The mode you want to run the robot in. Change it to reflect what you want to run it in. */
-    private static final RobotMode robot = RobotMode.ROBOT_2022_SIM;
+    // Enable or disable competition mode
+    public static final boolean competitionMode = false;
 
     // Joystick axis deadband for the swerve drive
     public static final double swerveDeadband = 0.1;
 
     // Value to voltage compensate the motors for
     public static final double voltageComp = 12.0;
-
-    /** Get the robot your code is running. */
-    public static RobotMode getRobot() {
-      if (RobotBase.isReal()) {
-        if (robot == RobotMode.ROBOT_2022_SIM) {
-          return RobotMode.ROBOT_2022_REAL;
-        } else {
-          return robot;
-        }
-      } else {
-        return robot;
-      }
-    }
-
-    /** Get the mode of what you're running the robot in. */
-    public static RunningMode getMode() {
-      switch (getRobot()) {
-        case ROBOT_2022_REAL:
-          return RobotBase.isReal() ? RunningMode.REAL : RunningMode.REPLAY;
-        case ROBOT_2022_SIM:
-          return RunningMode.SIM;
-        default:
-          return RunningMode.REAL;
-      }
-    }
   }
 
   /** Constants revolving around joysticks */
