+++ src/main/java/frc3512/robot/RobotContainer.java
@@ -9,12 +9,7 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc3512.robot.auton.TestAuto;
 import frc3512.robot.commands.driving.TeleopSwerve;
-import frc3512.robot.subsystems.drive.Swerve;
-import frc3512.robot.subsystems.drive.gyro.GyroIO;
-import frc3512.robot.subsystems.drive.gyro.GyroIOPigeon2;
-import frc3512.robot.subsystems.drive.module.ModuleIO;
-import frc3512.robot.subsystems.drive.module.ModuleIOSim;
-import frc3512.robot.subsystems.drive.module.ModuleIOSparkMAX;
+import frc3512.robot.subsystems.Swerve;
 
 /**
  * This class is where the bulk of the robot should be declared. Since Command-based is a
@@ -27,7 +22,7 @@ public class RobotContainer {
   private final SendableChooser<Command> m_autonChooser = new SendableChooser<Command>();
 
   // Robot subsystems
-  private Swerve m_swerve;
+  private Swerve m_swerve = new Swerve();
 
   // Xbox controllers
   private final Joystick driver = new Joystick(Constants.Joysticks.xboxControllerPort);
@@ -45,42 +40,6 @@ public class RobotContainer {
 
   /** The container for the robot. Contains subsystems, OI devices, and commands. */
   public RobotContainer() {
-    if (Constants.General.getMode() != Constants.RunningMode.REPLAY) {
-      switch (Constants.General.getRobot()) {
-        case ROBOT_2022_REAL:
-          m_swerve =
-              new Swerve(
-                  new GyroIOPigeon2(),
-                  new ModuleIOSparkMAX(0),
-                  new ModuleIOSparkMAX(1),
-                  new ModuleIOSparkMAX(2),
-                  new ModuleIOSparkMAX(3));
-          break;
-        case ROBOT_2022_SIM:
-          m_swerve =
-              new Swerve(
-                  new GyroIO() {},
-                  new ModuleIOSim(),
-                  new ModuleIOSim(),
-                  new ModuleIOSim(),
-                  new ModuleIOSim());
-          break;
-        default:
-          break;
-      }
-    }
-
-    // Register any missing subsystems
-    m_swerve =
-        m_swerve != null
-            ? m_swerve
-            : new Swerve(
-                new GyroIO() {},
-                new ModuleIO() {},
-                new ModuleIO() {},
-                new ModuleIO() {},
-                new ModuleIO() {});
-
     configureButtonBindings();
     configureAxisActions();
     registerAutons();
