+++ src/main/java/frc3512/robot/subsystems/SwerveModule.java
@@ -1,0 +1,166 @@
+package frc3512.robot.subsystems;
+
+import com.ctre.phoenix.sensors.CANCoder;
+import com.revrobotics.CANSparkMax;
+import com.revrobotics.CANSparkMax.ControlType;
+import com.revrobotics.CANSparkMaxLowLevel.MotorType;
+import com.revrobotics.RelativeEncoder;
+import com.revrobotics.SparkMaxPIDController;
+import edu.wpi.first.math.controller.SimpleMotorFeedforward;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.kinematics.SwerveModuleState;
+import frc3512.lib.config.SwerveModuleConstants;
+import frc3512.lib.math.OnboardModuleState;
+import frc3512.lib.util.CANCoderUtil;
+import frc3512.lib.util.CANCoderUtil.CANCoderUsage;
+import frc3512.lib.util.CANSparkMaxUtil;
+import frc3512.lib.util.CANSparkMaxUtil.Usage;
+import frc3512.robot.Constants;
+import frc3512.robot.Robot;
+
+public class SwerveModule {
+  public int moduleNumber;
+  private Rotation2d lastAngle;
+  private Rotation2d angleOffset;
+
+  private CANSparkMax angleMotor;
+  private CANSparkMax driveMotor;
+
+  private RelativeEncoder driveEncoder;
+  private RelativeEncoder integratedAngleEncoder;
+  private CANCoder angleEncoder;
+
+  private final SparkMaxPIDController driveController;
+  private final SparkMaxPIDController angleController;
+
+  private final SimpleMotorFeedforward feedforward =
+      new SimpleMotorFeedforward(
+          Constants.Swerve.driveKS, Constants.Swerve.driveKV, Constants.Swerve.driveKA);
+
+  /**
+   * Creates a new swerve module with NEO motors and a CTRE CANCoder.
+   *
+   * @param moduleNumber - Number of the module (0-3)
+   * @param moduleIO - Module IO for the appropriate module
+   */
+  public SwerveModule(int moduleNumber, SwerveModuleConstants moduleConstants) {
+    this.moduleNumber = moduleNumber;
+    angleOffset = moduleConstants.angleOffset;
+
+    /* Angle Encoder Config */
+    angleEncoder = new CANCoder(moduleConstants.cancoderID);
+    configAngleEncoder();
+
+    /* Angle Motor Config */
+    angleMotor = new CANSparkMax(moduleConstants.angleMotorID, MotorType.kBrushless);
+    integratedAngleEncoder = angleMotor.getEncoder();
+    angleController = angleMotor.getPIDController();
+    configAngleMotor();
+
+    /* Drive Motor Config */
+    driveMotor = new CANSparkMax(moduleConstants.driveMotorID, MotorType.kBrushless);
+    driveEncoder = driveMotor.getEncoder();
+    driveController = driveMotor.getPIDController();
+    configDriveMotor();
+
+    lastAngle = getState().angle;
+  }
+
+  private void configAngleEncoder() {
+    angleEncoder.configFactoryDefault();
+    CANCoderUtil.setCANCoderBusUsage(angleEncoder, CANCoderUsage.kMinimal);
+    angleEncoder.configAllSettings(Robot.ctreConfigs.swerveCanCoderConfig);
+  }
+
+  private void configAngleMotor() {
+    angleMotor.restoreFactoryDefaults();
+    CANSparkMaxUtil.setCANSparkMaxBusUsage(angleMotor, Usage.kPositionOnly);
+    angleMotor.setSmartCurrentLimit(Constants.Swerve.angleContinuousCurrentLimit);
+    angleMotor.setInverted(Constants.Swerve.angleInvert);
+    angleMotor.setIdleMode(Constants.Swerve.angleNeutralMode);
+    integratedAngleEncoder.setPositionConversionFactor(Constants.Swerve.angleConversionFactor);
+    angleController.setP(Constants.Swerve.angleKP);
+    angleController.setI(Constants.Swerve.angleKI);
+    angleController.setD(Constants.Swerve.angleKD);
+    angleController.setFF(Constants.Swerve.angleKFF);
+    angleMotor.enableVoltageCompensation(Constants.General.voltageComp);
+    angleMotor.burnFlash();
+    resetAbsolute();
+  }
+
+  private void configDriveMotor() {
+    driveMotor.restoreFactoryDefaults();
+    CANSparkMaxUtil.setCANSparkMaxBusUsage(driveMotor, Usage.kAll);
+    driveMotor.setSmartCurrentLimit(Constants.Swerve.driveContinuousCurrentLimit);
+    driveMotor.setInverted(Constants.Swerve.driveInvert);
+    driveMotor.setIdleMode(Constants.Swerve.driveNeutralMode);
+    driveEncoder.setVelocityConversionFactor(Constants.Swerve.driveConversionVelocityFactor);
+    driveEncoder.setPositionConversionFactor(Constants.Swerve.driveConversionPositionFactor);
+    driveController.setP(Constants.Swerve.angleKP);
+    driveController.setI(Constants.Swerve.angleKI);
+    driveController.setD(Constants.Swerve.angleKD);
+    driveController.setFF(Constants.Swerve.angleKFF);
+    driveMotor.enableVoltageCompensation(Constants.General.voltageComp);
+    driveMotor.burnFlash();
+    driveEncoder.setPosition(0.0);
+  }
+
+  private void resetAbsolute() {
+    double absolutePosition = getCanCoder().getDegrees() - angleOffset.getDegrees();
+    integratedAngleEncoder.setPosition(absolutePosition);
+  }
+
+  public void setDesiredState(SwerveModuleState desiredState, boolean isOpenLoop) {
+    // Custom optimize command, since default WPILib optimize assumes continuous controller which
+    // REV and CTRE are not
+    desiredState = OnboardModuleState.optimize(desiredState, getState().angle);
+
+    setAngle(desiredState);
+    setSpeed(desiredState, isOpenLoop);
+  }
+
+  private void setSpeed(SwerveModuleState desiredState, boolean isOpenLoop) {
+    if (isOpenLoop) {
+      double percentOutput = desiredState.speedMetersPerSecond / Constants.Swerve.maxSpeed;
+      driveMotor.set(percentOutput);
+    } else {
+      driveController.setReference(
+          desiredState.speedMetersPerSecond,
+          ControlType.kVelocity,
+          0,
+          feedforward.calculate(desiredState.speedMetersPerSecond));
+    }
+  }
+
+  private void setAngle(SwerveModuleState desiredState) {
+    // Prevent rotating module if speed is less then 1%. Prevents jittering.
+    Rotation2d angle =
+        (Math.abs(desiredState.speedMetersPerSecond) <= (Constants.Swerve.maxSpeed * 0.01))
+            ? lastAngle
+            : desiredState.angle;
+
+    angleController.setReference(angle.getDegrees(), ControlType.kPosition);
+    lastAngle = angle;
+  }
+
+  private Rotation2d getAngle() {
+    return Rotation2d.fromDegrees(integratedAngleEncoder.getPosition());
+  }
+
+  public Rotation2d getCanCoder() {
+    return Rotation2d.fromDegrees(angleEncoder.getAbsolutePosition());
+  }
+
+  public void stop() {
+    driveMotor.stopMotor();
+    angleMotor.stopMotor();
+  }
+
+  public SwerveModuleState getState() {
+    return new SwerveModuleState(driveEncoder.getVelocity(), getAngle());
+  }
+
+  public void periodic() {
+
+  }
+}
