--- vendordeps/AdvantageKit.json
@@ -1,41 +1,0 @@
-{
-    "fileName": "AdvantageKit.json",
-    "name": "AdvantageKit",
-    "version": "1.8.1",
-    "uuid": "d820cc26-74e3-11ec-90d6-0242ac120003",
-    "mavenUrls": [],
-    "jsonUrl": "https://github.com/Mechanical-Advantage/AdvantageKit/releases/latest/download/AdvantageKit.json",
-    "javaDependencies": [
-        {
-            "groupId": "org.littletonrobotics.akit.junction",
-            "artifactId": "wpilib-shim",
-            "version": "1.8.1"
-        },
-        {
-            "groupId": "org.littletonrobotics.akit.junction",
-            "artifactId": "junction-core",
-            "version": "1.8.1"
-        },
-        {
-            "groupId": "org.littletonrobotics.akit.conduit",
-            "artifactId": "conduit-api",
-            "version": "1.8.1"
-        }
-    ],
-    "jniDependencies": [
-        {
-            "groupId": "org.littletonrobotics.akit.conduit",
-            "artifactId": "conduit-wpilibio",
-            "version": "1.8.1",
-            "skipInvalidPlatforms": true,
-            "isJar": false,
-            "validPlatforms": [
-                "linuxathena",
-                "windowsx86-64",
-                "linuxx86-64",
-                "osxx86-64"
-            ]
-        }
-    ],
-    "cppDependencies": []
-}
